refactor: skill list
