+++ src/features/SkillStore/SkillList/LobeHub/index.tsx
@@ -4,6 +4,10 @@ import { KLAVIS_SERVER_TYPES, LOBEHUB_SKILL_PROVIDERS } from '@lobechat/const';
 import isEqual from 'fast-deep-equal';
 import { memo, useCallback, useMemo } from 'react';
 
+import {
+  createKlavisSkillDetailModal,
+  createLobehubSkillDetailModal,
+} from '@/features/SkillStore/SkillDetail';
 import { serverConfigSelectors, useServerConfigStore } from '@/store/serverConfig';
 import { useToolStore } from '@/store/tool';
 import { klavisStoreSelectors, lobehubSkillStoreSelectors } from '@/store/tool/selectors';
@@ -11,7 +15,6 @@ import { KlavisServerStatus } from '@/store/tool/slices/klavisStore';
 import { LobehubSkillStatus } from '@/store/tool/slices/lobehubSkillStore/types';
 
 import Empty from '../Empty';
-import { createKlavisSkillDetailModal, createLobehubSkillDetailModal } from '../SkillDetail';
 import { gridStyles } from '../style';
 import Item from './Item';
